requirements.txt
pillow
markdown

django
django-crispy-forms
crispy-bootstrap5
django-browser-reload

djangorestframework
markdown
django-filter
